third_party/blink/web_tests/virtual/select-parser-relaxation/external/wpt/html/semantics/forms/the-select-element/stylable-select/option-computed-style.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] appearance:base-select options should have a checkmark with empty alt text.
  assert_equals: expected "\\"✓\\" / \\"\\"" but got "none"
Harness: the test ran to completion.
